// apps/api/src/application/common/ErrorCodes.ts
export enum ErrorCodes {
  VALIDATION_ERROR = "VALIDATION_ERROR",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  SESSION_NOT_FOUND = "SESSION_NOT_FOUND",
  FLOW_NOT_FOUND = "FLOW_NOT_FOUND",
  FLOW_ALREADY_EXISTS = "FLOW_ALREADY_EXISTS",
  FLOW_ID_MISMATCH = "FLOW_ID_MISMATCH",
  DUPLICATE_NODE_ID = "DUPLICATE_NODE_ID",
  EMPTY_FLOW = "EMPTY_FLOW",
  INVALID_START_NODE = "INVALID_START_NODE",
  MISSING_START_NODE = "MISSING_START_NODE",
  MULTIPLE_START_NODES = "MULTIPLE_START_NODES",
  DUPLICATE_EDGE_ID = "DUPLICATE_EDGE_ID",
  EDGE_SOURCE_NOT_FOUND = "EDGE_SOURCE_NOT_FOUND",
  EDGE_TARGET_NOT_FOUND = "EDGE_TARGET_NOT_FOUND",
  MISSING_RESPONSE_NODE = "MISSING_RESPONSE_NODE",
  UNREACHABLE_NODE = "UNREACHABLE_NODE"
}